+++ openeo/rest/connection.py
@@ -719,9 +719,13 @@ class Connection(RestApiConnection):
             self, user_defined_process_id: str,
             process_graph: Union[dict, ProcessBuilderBase],
             parameters: List[Union[dict, Parameter]] = None,
-            public: bool = False, summary: str = None, description: str = None,
-            returns: Optional[dict] = None, categories: Optional[List[str]] = None, 
-            examples: Optional[List[dict]] = None, links: Optional[List[dict]] = None
+            public: bool = False,
+            summary: Optional[str] = None,
+            description: Optional[str] = None,
+            returns: Optional[dict] = None,
+            categories: Optional[List[str]] = None,
+            examples: Optional[List[dict]] = None,
+            links: Optional[List[dict]] = None,
     ) -> RESTUserDefinedProcess:
         """
         Store a process graph and its metadata on the backend as a user-defined process for the authenticated user.
@@ -745,8 +749,10 @@ class Connection(RestApiConnection):
             warnings.warn("Defining user-defined process {u!r} without parameters".format(u=user_defined_process_id))
         udp = RESTUserDefinedProcess(user_defined_process_id=user_defined_process_id, connection=self)
         udp.store(
-            process_graph=process_graph, parameters=parameters, public=public, summary=summary, description=description,
-            returns=returns, categories=categories, examples=examples, links=links)
+            process_graph=process_graph, parameters=parameters, public=public,
+            summary=summary, description=description,
+            returns=returns, categories=categories, examples=examples, links=links
+        )
         return udp
 
     def list_user_defined_processes(self) -> List[dict]: